Renamed and converted to HasAttr object
